add parsing for Card data

+++ OpenNos.DAL.EF/Entities/Card.cs
@@ -39,7 +39,7 @@ namespace OpenNos.DAL.EF
 
         public int EffectId { get; set; }
 
-        public short FirstData { get; set; }
+        public int FirstData { get; set; }
 
         public byte Level { get; set; }
 
@@ -50,13 +50,13 @@ namespace OpenNos.DAL.EF
 
         public byte Propability { get; set; }
 
-        public short SecondData { get; set; }
+        public int SecondData { get; set; }
 
         public virtual ICollection<SkillCard> SkillCard { get; set; }
 
         public byte SubType { get; set; }
 
-        public byte Type { get; set; }
+        public short Type { get; set; }
 
         #endregion
     }
